Create 15_schoolchildren.java
An interesting task. It was necessary to make a program that generated exercises for schoolchildren. It was said in the condition that it is not necessary to display examples with the numbers 1 and 10 and the same type of calculations, for example, we display 3*2=6 and 2*3=6, it is no longer necessary to display